requirements.txt
Flask
flask-cors
python-dotenv
google-api-python-client
nltk
gunicorn